+++ os/ex/ST/lis302dl.c
@@ -140,7 +140,7 @@ static msg_t read_cooked(void *ip, float axes[]) {
   return msg;
 }
 
-static msg_t set_bias(void *ip, int32_t *bp) {
+static msg_t set_bias(void *ip, float *bp) {
   uint32_t i;
   
   osalDbgCheck((ip != NULL) && (bp !=NULL));
